refactor(tasks): extract task list data hook
